Module feed_rs::model

Structs

Category
Represents the category of a feed or entry

Content
Content, or link to the content, for a given entry.

Entry
An item within a feed

Feed
Combined model for a syndication feed (i.e. RSS1, RSS 2, Atom, JSON Feed)

Generator
Information on the tools used to generate the feed

Image
Represents a a link to an image.

Link
Represents a link to an associated resource for the feed or entry.

Person
Represents an author, contributor etc.

Text
Textual content, or link to the content, for a given entry.